+++ src/remindly/__init__.py
@@ -1,3 +1,3 @@
 """Remindly."""
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"